Second Homework Submission
I have been through some hardships in troublshooting and trying to figure out why some css is not reflecting in in the html related to it specially the vertical centering. It took me five hours to reach this result however not all of it was coding it was mostly researching how to reach such output.